# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API reto Torneo Solidario",
        "description": "API reto Torneo Solidario",
        "contact": {
            "email": "email@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/donaciones": {
            "get": {
                "tags": [
                    "donaciones"
                ],
                "summary": "Listado de donaciones",
                "description": "Listado de todas las donaciones",
                "operationId": "getDonaciones",
                "responses": {
                    "200": {
                        "description": "Listado de donaciones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/donaciones"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "donaciones"
                ],
                "summary": "Crear una donación",
                "description": "Crear una nueva donación",
                "operationId": "postDonacion",
                "requestBody": {
                    "description": "Datos de la donación",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ong_id",
                                    "kilos",
                                    "importe"
                                ],
                                "properties": {
                                    "ong_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "kilos": {
                                        "type": "integer",
                                        "example": "10"
                                    },
                                    "importe": {
                                        "type": "integer",
                                        "example": "100"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Donación creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/donaciones"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos no válidos"
                    }
                }
            }
        },
        "/api/donaciones/{id}": {
            "get": {
                "tags": [
                    "donaciones"
                ],
                "summary": "Obtener una donación",
                "description": "Obtener una donación por su id",
                "operationId": "getDonacion",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del producto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Donación encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/donaciones"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Donación no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "donaciones"
                ],
                "summary": "Actualizar una donación",
                "description": "Actualizar una donación por su id",
                "operationId": "updateDonacion",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de la donación",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos de la donación",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ong_id",
                                    "kilos",
                                    "importe"
                                ],
                                "properties": {
                                    "ong_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "kilos": {
                                        "type": "integer",
                                        "example": "10"
                                    },
                                    "importe": {
                                        "type": "integer",
                                        "example": "100"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Donación actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/donaciones"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Donación no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "donaciones"
                ],
                "summary": "Eliminar una donación",
                "description": "Eliminar una donación por su id",
                "operationId": "deleteDonacion",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de la donación",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Donación eliminada"
                    },
                    "404": {
                        "description": "Donación no encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "donaciones": {
                "title": "Donaciones",
                "properties": {
                    "ong_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "kilos": {
                        "type": "integer",
                        "example": "10"
                    },
                    "importe": {
                        "type": "integer",
                        "example": "100"
                    },
                    "usuarioIdCreacion": {
                        "type": "integer",
                        "example": "1"
                    },
                    "fechaCreacion": {
                        "type": "timestamp",
                        "example": "2022-02-11 15:12:24"
                    },
                    "usuarioIdActualizacion": {
                        "type": "integer",
                        "example": "1"
                    },
                    "fechaActualizacion": {
                        "type": "timestamp",
                        "example": "2022-02-11 15:12:24"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}